+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## v0.9.5
 
-- let Main CI use the correct Node.js version
-- add Black-box Tests
-- add CI Matrix
+- upgrade Node.js runtime to v7.7.2
+- let Master CI use the correct Node.js version
+- add Black-box Test
+- add RAM Test
 - make sure that the user have installed the correct version of node in her environment; it should match the enclosed Node.js runtime version of the compiler
 
 ## v0.9.4